Assembler moved to core
