+++ documentation/src/main/cheatsheet/cheatsheets.adoc
@@ -146,6 +146,38 @@ Set the root folder of the repository on file system.
  If catalogue equals empty string a verticle will look for the files in classpath
 +++
 |===
+[[FragmentAssemblerOptions]]
+== FragmentAssemblerOptions
+
+++++
+ Describes FragmentAssembler Knot configuration
+++++
+'''
+
+[cols=">25%,^25%,50%"]
+[frame="topbot"]
+|===
+^|Name | Type ^| Description
+|[[address]]`address`|`String`|
++++
+Set the EB address of the verticle
++++
+|[[snippetTagName]]`snippetTagName`|`String`|
++++
+Sets a Knot.x snippet HTML tag name. Default is 'script'
++++
+|[[unprocessedStrategy]]`unprocessedStrategy`|`link:cheatsheets.adoc#UnprocessedFragmentStrategy[UnprocessedFragmentStrategy]`|
++++
+Set the strategy how to assembly markup with snippets that were not processed by any Knot.
+ Allowed values are:
+ <ul>
+ <li>AS_IS - Keep the whole unprocessed snippet as is</li>
+ <li>UNWRAP - Remove the wrapping script tag from the snippet</li>
+ <li>IGNORE - Remove snippet from the markup</li>
+ </ul>
+ If not set, a default value is <b>UNWRAP</b>
++++
+|===
 [[FragmentSplitterOptions]]
 == FragmentSplitterOptions
