+++ documentation/src/main/wiki/KnotxDeployment.md
@@ -74,7 +74,7 @@ As mentioned above, the `knotx-starter.json` is the main configuration file desc
     "httpRepo=io.knotx.repository.http.HttpRepositoryConnectorVerticle",
     "fsRepo=io.knotx.repository.fs.FilesystemRepositoryConnectorVerticle",
     "splitter=io.knotx.splitter.FragmentSplitterVerticle",
-    "assembler=io.knotx.knot.assembler.FragmentAssemblerVerticle",
+    "assembler=io.knotx.assembler.FragmentAssemblerVerticle",
     "hbsKnot=io.knotx.knot.templating.HandlebarsKnotVerticle",
     "serviceKnot=io.knotx.knot.service.ServiceKnotVerticle",
     "actionKnot=io.knotx.knot.action.ActionKnotVerticle",
@@ -103,7 +103,7 @@ For instance, if you want to modify configuration of KnotxServer module, you can
     "httpRepo=io.knotx.repository.http.HttpRepositoryConnectorVerticle",
     "fsRepo=io.knotx.repository.fs.FilesystemRepositoryConnectorVerticle",
     "splitter=io.knotx.splitter.FragmentSplitterVerticle",
-    "assembler=io.knotx.knot.assembler.FragmentAssemblerVerticle",
+    "assembler=io.knotx.assembler.FragmentAssemblerVerticle",
     "hbsKnot=io.knotx.knot.templating.HandlebarsKnotVerticle",
     "serviceKnot=io.knotx.knot.service.ServiceKnotVerticle",
     "actionKnot=io.knotx.knot.action.ActionKnotVerticle",
@@ -163,7 +163,7 @@ and the list of deployment status at the end shows:
 
 		Failed deploying customKnot [java:com.acme.MyCustomModuleVerticle]
 		Deployed fsRepo [java:io.knotx.repository.fs.FilesystemRepositoryConnectorVerticle] [ef35d8e8-887b-4f8b-bd2e-31943d8ff706]
-		Deployed assembler [java:io.knotx.knot.assembler.FragmentAssemblerVerticle] [0f5d74f4-2233-4937-b11c-d20d3e5451b1]
+		Deployed assembler [java:io.knotx.assembler.FragmentAssemblerVerticle] [0f5d74f4-2233-4937-b11c-d20d3e5451b1]
 		...
 ```
