--- knotx-knot/knotx-knot-fragment-assembler/src/main/resources/codegen.json
@@ -1,14 +1,0 @@
-{
-  "name": "knotx-cheatsheet",
-  "generators": [{
-    "kind": "dataObject",
-    "incremental": true,
-    "filename": "'../../../../documentation/.dataobjects/fragment-assembler-dataobjects.adoc'",
-    "templateFilename": "knotx-codegen/dataObjectCheatsheet.templ"
-  },{
-    "kind": "enum",
-    "incremental": true,
-    "filename": "'../../../documentation/.dataobjects/fragment-assembler-enums-dataobjects.adoc'",
-    "templateFilename": "vertx-java/template/enumCheatsheet.templ"
-  }]
-}
