--- knotx-knot/knotx-knot-fragment-assembler/src/test/resources/test.unwrap.io.knotx.FragmentAssembler.json
@@ -1,5 +1,0 @@
-{
-  "modules": [
-    "assembler=io.knotx.knot.assembler.FragmentAssemblerVerticle"
-  ]
-}
