+++ knotx-standalone/config/application.conf
@@ -3,7 +3,7 @@ modules = [
     "httpRepo=io.knotx.repository.http.HttpRepositoryConnectorVerticle"
     "fsRepo=io.knotx.repository.fs.FilesystemRepositoryConnectorVerticle"
     "splitter=io.knotx.splitter.FragmentSplitterVerticle"
-    "assembler=io.knotx.knot.assembler.FragmentAssemblerVerticle"
+    "assembler=io.knotx.assembler.FragmentAssemblerVerticle"
     "hbsKnot=io.knotx.knot.templating.HandlebarsKnotVerticle"
     "serviceKnot=io.knotx.knot.service.ServiceKnotVerticle"
     "actionKnot=io.knotx.knot.action.ActionKnotVerticle"
