+++ knotx-standalone/pom.xml
@@ -39,11 +39,6 @@
     </dependency>
     <dependency>
       <groupId>io.knotx</groupId>
-      <artifactId>knotx-knot-fragment-assembler</artifactId>
-      <version>${project.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>io.knotx</groupId>
       <artifactId>knotx-knot-service</artifactId>
       <version>${project.version}</version>
     </dependency>
