+++ tmp-core/src/main/java/io/knotx/assembler/FragmentAssemblerOptions.java
@@ -13,9 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.knotx.knot.assembler;
+package io.knotx.assembler;
 
-import io.knotx.knot.assembler.impl.UnprocessedFragmentStrategy;
 import io.vertx.codegen.annotations.DataObject;
 import io.vertx.core.json.JsonObject;
