+++ tmp-core/src/test/resources/big_snippet.txt
