+++ tmp-core/src/test/resources/customTag_snippet.txt
