+++ tmp-core/src/test/resources/expectedAsIsResult.html
