+++ tmp-core/src/test/resources/fragment2.txt
