+++ tmp-core/src/test/resources/fragment3.txt
