+++ tmp-core/src/test/resources/raw_fragment.txt
