+++ tmp-core/src/test/resources/simple_snippet-expected_ignored_content.txt
