+++ tmp-core/src/test/resources/simple_snippet.txt
