+++ tmp-core/src/test/resources/test.asIs.io.knotx.FragmentAssembler.json
@@ -1,12 +1,12 @@
 {
   "modules": [
-    "assembler=io.knotx.knot.assembler.FragmentAssemblerVerticle"
+    "assembler=io.knotx.assembler.FragmentAssemblerVerticle"
   ],
   "config": {
     "assembler": {
       "options": {
         "config": {
-          "unprocessedStrategy": "IGNORE"
+          "unprocessedStrategy": "AS_IS"
         }
       }
     }
