+++ tmp-core/src/test/resources/test.unwrap.io.knotx.FragmentAssembler.json
@@ -1,0 +1,5 @@
+{
+  "modules": [
+    "assembler=io.knotx.assembler.FragmentAssemblerVerticle"
+  ]
+}
